+++ fs/tfs_server.c
@@ -12,25 +12,36 @@
 static char active_sessions[S][PIPE_NAME_LENGTH];
 static bool shutdown;
 
-pthread_mutex_t session_lock = PTHREAD_MUTEX_INITIALIZER;
+static pthread_mutex_t session_lock = PTHREAD_MUTEX_INITIALIZER;
 
-// TODO: comment
+/*
+ * Activates a session with a client (if it is possible).
+ *
+ * Returns the session_id where the session is created if successful, 
+ * -1 otherwise.
+ */
 int tfs_mount(char const *client_pipe_name){
 	/* Only one client shall mount/unmount at one time */
 	if(pthread_mutex_lock(&session_lock) != 0)
 		return -1;
-
-	if(active_sessions[0] == 0){
+	
+	int ret = -1;
+	if(strlen(active_sessions[0]) == 0){
 		memcpy(active_sessions[0], client_pipe_name, PIPE_NAME_LENGTH);
+		ret = 0;
 	}
 
 	if(pthread_mutex_unlock(&session_lock) != 0)
 		return -1;
-	return 0;
+	return ret;
 	/* return session_id */
 }
 
-// TODO: comment
+/* 
+ * Removes the session active on session_id.
+ *
+ * Returns 0 if successful, -1 otherwise.
+ */
 int tfs_unmount(int session_id){
 	/* Only one client shall mount/unmount at one time */
 	if(pthread_mutex_lock(&session_lock) != 0)
@@ -43,16 +54,18 @@ int tfs_unmount(int session_id){
 	return 0;
 }
 
+/* Returns the value of tfs_mount on success, -1 otherwise */
 int process_mount(int fserv){
 	/* get the client pipe name */
 	char client_pipe_name[PIPE_NAME_LENGTH];
 	ssize_t rd = read(fserv, client_pipe_name, PIPE_NAME_LENGTH*sizeof(char));
-	if (rd == 0)
+	if (rd == 0){
 		// TODO: return what? should we open/close&open?
 		return 0;
-	else if (rd == -1)
+	}else if (rd == -1){
 		// TODO: should we use exit(EXIT_FAILURE) or return -1
 		return -1;
+	}
 
 	int ret = tfs_mount(client_pipe_name);
 
@@ -61,20 +74,23 @@ int process_mount(int fserv){
 	if(fcli == -1)
 		return -1;
 
-	write(fcli, &ret, sizeof(int));
+	if(write(fcli, &ret, sizeof(int)) == -1)
+		return -1;
 	return ret;
 }
 
+/* Returns the value of tfs_unmount on success, -1 otherwise */
 int process_unmount(int session_id, int fcli){
 	int ret = tfs_unmount(session_id);
-	write(fcli, &ret, sizeof(int));
+	if(write(fcli, &ret, sizeof(int)) == -1)
+		return -1;
 	return ret;
 }
 
+/* Returns the value of tfs_open on success, -1 otherwise */
 int process_open(int fserv, int fcli){
 	char name[PIPE_NAME_LENGTH];
 	ssize_t rd = read(fserv, name, PIPE_NAME_LENGTH*sizeof(char));
-	// TODO
 	if (rd == 0)
 		return 0;
 	else if (rd == -1)
@@ -82,35 +98,36 @@ int process_open(int fserv, int fcli){
 
 	int flags;
 	rd = read(fserv, &flags, sizeof(int));
-	// TODO
 	if (rd == 0)
 		return 0;
 	else if (rd == -1)
 		return -1;
 
 	int ret = tfs_open(name, flags);
-	write(fcli, &ret, sizeof(int));
+	if(write(fcli, &ret, sizeof(int)) == -1)
+		return -1;
 	return ret;
 }
 
+/* Returns the value of tfs_close on success, -1 otherwise */
 int process_close(int fserv, int fcli){
 	int fhandle;
 	ssize_t rd = read(fserv, &fhandle, sizeof(int));
-	// TODO
 	if (rd == 0)
 		return 0;
 	else if (rd == -1)
 		return -1;
 
 	int ret = tfs_close(fhandle);
-	write(fcli, &ret, sizeof(int));
+	if(write(fcli, &ret, sizeof(int)) == -1)
+		return -1;
 	return ret;
 }
 
+/* Returns the value of tfs_write on success, -1 otherwise */
 int process_write(int fserv, int fcli){
 	int fhandle;
 	ssize_t rd = read(fserv, &fhandle, sizeof(int));
-	// TODO
 	if (rd == 0)
 		return 0;
 	else if (rd == -1)
@@ -118,7 +135,6 @@ int process_write(int fserv, int fcli){
 
 	size_t len;
 	rd = read(fserv, &len, sizeof(size_t));
-	// TODO
 	if (rd == 0)
 		return 0;
 	else if (rd == -1)
@@ -126,21 +142,21 @@ int process_write(int fserv, int fcli){
 
 	char content[len];
 	rd = read(fserv, content, len*sizeof(char));
-	// TODO
 	if (rd == 0)
 		return 0;
 	else if (rd == -1)
 		return -1;
 
 	int ret = (int) tfs_write(fhandle, content, len);
-	write(fcli, &ret, sizeof(int));
+	if(write(fcli, &ret, sizeof(int)) == -1)
+		return -1;
 	return ret;
 }
 
+/* Returns the value of tfs_read on success, -1 otherwise */
 int process_read(int fserv, int fcli){
 	int fhandle;
 	ssize_t rd = read(fserv, &fhandle, sizeof(int));
-	// TODO
 	if (rd == 0)
 		return 0;
 	else if (rd == -1)
@@ -148,7 +164,6 @@ int process_read(int fserv, int fcli){
 
 	size_t len;
 	rd = read(fserv, &len, sizeof(size_t));
-	// TODO
 	if (rd == 0)
 		return 0;
 	else if (rd == -1)
@@ -157,41 +172,43 @@ int process_read(int fserv, int fcli){
 	char buf[len];
 
 	int ret = (int)tfs_read(fhandle, buf, len);
-	write(fcli, &ret, sizeof(int));
+	if(write(fcli, &ret, sizeof(int)) == -1)
+		return -1;
 
 	if(ret != -1)
-		write(fcli, buf, (size_t)ret*sizeof(char));
+		if(write(fcli, buf, (size_t)ret*sizeof(char)) == -1)
+			return -1;
 	return ret;
 }
 
+/* Returns the value of tfs_destroy_after_all_closed on success, -1 otherwise */
 int process_shutdown_aac(int fcli){
 	int ret = tfs_destroy_after_all_closed();
 	if(ret == 0)
 		shutdown = 0;
-	write(fcli, &ret, sizeof(int));
+	if(write(fcli, &ret, sizeof(int)) == -1)
+		return -1;
 	return ret;
 }
 
+/* Returns the value of the funcion requested by the client if successful, 
+ * -1 otherwise */
 int process_message(int fserv){
 	/* get op code */
 	char op_code;
 	ssize_t rd = read(fserv, &op_code, sizeof(char));
-	// TODO
 	if (rd == 0)
 		return 0;
 	else if (rd == -1)
 		return -1;
 
 	if(op_code == 1){
-		process_mount(fserv);
-		return 0;
+		return process_mount(fserv);
 	}
 	/* else */
-
 	/* get session id */
 	int session_id;
 	rd = read(fserv, &session_id, sizeof(int));
-	// TODO
 	if (rd == 0)
 		return 0;
 	else if (rd == -1)
@@ -204,35 +221,33 @@ int process_message(int fserv){
 		return -1;
 
 	switch(op_code){
-		// TODO: tentar mudar para os codes como deve ser
 		case 2:
-			process_unmount(session_id, fcli);
+			return process_unmount(session_id, fcli);
 			break;
 
 		case 3:
-			process_open(fserv, fcli);
+			return process_open(fserv, fcli);
 			break;
 
 		case 4:
-			process_close(fserv, fcli);
+			return process_close(fserv, fcli);
 			break;
 
 		case 5:
-			process_write(fserv, fcli);
+			return process_write(fserv, fcli);
 			break;
 		
 		case 6:
-			process_read(fserv, fcli);
+			return process_read(fserv, fcli);
 			break;
 		
 		case 7:
-			process_shutdown_aac(fcli);
+			return process_shutdown_aac(fcli);
 			break;
 
 		default:
 	}
-
-	return 0;
+	return -1;
 }
 
 int main(int argc, char **argv) {
@@ -246,11 +261,14 @@ int main(int argc, char **argv) {
 
 	/* creating server pipe and opening for read */
 	unlink(pipename);
+
 	if(mkfifo(pipename, 0777) < 0)
 		exit(1);
 
 	int fserv = open(pipename, O_RDONLY);
 	if(fserv < 0) exit(1);
+
+	memset(active_sessions, '\0', S*PIPE_NAME_LENGTH);
 
 	if(tfs_init() == -1)
 		return -1;
@@ -260,13 +278,12 @@ int main(int argc, char **argv) {
 		process_message(fserv);
 	}
 
+	if(close(fserv) != 0) 
+		return -1;
+
 	/* Destruction process */
 	unlink(pipename);
 
+
     return 0;
 }
-
-// TODO: check locks/unlocks work
-// TODO: implement thread beahviour for different sessions
-// TODO: o que acontece se o servidor não conseguir abrir o pipe do cliente?
-// TODO: check returns in process_ functions
